+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Static components now have acf_fc_layout field as well
+
+### Changed
+- Static components don't have a prefix anymore as this would break component data functions
+- Changed clonable static component name from c to component name
+
 ## [0.5.0] - 2017-09-08
 
 ### Changed